Tidied up services and policies to properly use Wisper, transactions and locking
Also more strongly defined a finished round:
A round model instance is finished if there exists a round after it.
A round is considered finished (and a new one should be created) if it meets some conditions in the
RoundFinishedPolicy.

Also refactored a few variables:
Renamed 'current_player' to 'active_player'.
Renamed 'current_user_player' to 'current_player'.